mcp-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
PyJWT==2.8.0
cryptography==41.0.8
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
PyYAML==6.0.1
